requirements.txt
fastapi
uvicorn
pylint
black
openai
python-dotenv
requests
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
